// frontend/src/api/generated/model/saveFarmViewModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { GeoCoordinateViewModel } from './geoCoordinateViewModel';

export interface SaveFarmViewModel {
  name: string;
  city: string;
  state: string;
  headquarters: null | GeoCoordinateViewModel;
}